pyrepackager: move Python bytecode compiling into pyrepackager crate
Initially, I put this in pyembed/build.rs because I wanted pyrepackager
to avoid a dependency on libpython. I think it is acceptable for this
build-time crate to have a dependency on libpython. There is potential
for version skew between the libpython used by pyrepackager and pyembed.
But since we only target Python 3.7 currently, we can worry about this
problem later.