pyrepackager: support linking a custom libpython
Previously, the build pulled in libpython using the logic in the
cpython/python3-sys crates. This commit teaches the build to link its
own libpython from the object files contained in the configured Python
distribution.

This commit breaks the build because Cargo shares features between
target and build modes. To work around, we'll need to introduce a
fork of the cpython/python3-sys crates. That will be done in the
next commit...