third_party: create copy of rust-cpython project
Cargo doesn't support passing differing crate features when building
build scripts from targets. See
https://github.com/rust-lang/cargo/issues/5730.

We require different features for the cpython crate. Specifically, we
want to force the cpython crate to not link against a libpython, instead
providing a custom libpython compiled from object files.

This Cargo limitation thus prevents us from building the cpython crate
the way we desire.

A workaround to this limitation is to create a copy of a crate with
a different name.

This commit creates a copy of the rust-cpython directory, which contains
the cpython and python3-sys crates. We modify the crates to have a new name
and to recognize the new names within crate features.

The pyrepackager crate is modified to use the copied crate.

After this change, pyrepackage is linked against a found Python 3.7
install. The pyembed crate links against the libpython generated from its
build.rs. And pyembed - being the user-facing crate - uses the unmodified
crates.

This is all very hacky. But until Cargo supports nice things, it appears
we have to resort to ugliness like this :/