pyrepackager: extract Python distribution to temporary directory
Storing contents in memory is excessive. And it precludes our ability to
write common functions that walk a filesystem and turn files into common
data structures.

This commit starts the process of moving distribution parsing to be
based on iterating files on the filesystem. The current implementation
is pretty bad because it extracts files to the filesystem and still
stores file data in memory buffers. This will be cleaned up in
subsequent commits.