pyrepackager: start to parse PYTHON.json files
python-build-standalone is now producing JSON files that describe the
distribution. This should make it possible to remove a lot of
the "parsing" logic from pyrepackager.

This commit teaches pyrepackager about PYTHON.json files.
We only parse them and use its value to find the Python standard
library, allowing us to remove the pkgconfig parsing code.

There is still a bit of ways to go. This commit effectively breaks
PyOxidizer.