rust-cpython: add feature to force static Python linking
The build script for python3-sys assumes that Python is being
provided by a shared library on Windows. This is obviously not
true for PyOxidizer!

This commit adds a crate feature to force the use of a static
libpython.

On Windows, we have to use the "static-nobundle" link type
to force static linking (to avoid name dllimport name mangling)
while also telling Rust that it is OK for the Python library to
not exist yet. This requires a Nightly Rust compiler, which is
unfortunate.

We /might/ be able to work around this. But it took me several
hours to figure out what's going on here and I don't want to sink
more time into it now. We can investigate a workaround later.