pyembed: use __acrt_iob_func to obtain stdin FILE* on Windows
fdopen(0) was crashing for unknown reasons. Looking at the MSCRT
headers, {stdin, stdout, stderr} are #define's that use the
internal CRT function __acrt_iob_func(). If we define this as an
external function and call it, things seem to "just work."

With this change, the Python REPL now works on Windows and we
have a single file executable containing a full Python
distribution!!!