pyembed: use env::args_os() for setting sys.argv
env::args() validates that arguments are Unicode. We may have non-Unicode
values, especially on POSIX, which only guarantees arguments are char*.

This commit switches things to env::args_os(), which exposes OsString
instances, from which we can get the raw values from the OS and cast
them to a Python str using an appropriate mechanism.

See #10 for more.