pyrepackager: support excluding module sources
Packaging sources in addition to bytecode is convenient because
it facilitates debugging. But it isn't strictly required, especially
if the program isn't expected to be debugged.

This commit implements config options to exclude sources from
the packaged resources.

On a binary I'm using for testing, this reduces the size of the
binary by ~30 MB!