pyoxidizer: introduce project
The goal is to have a `pyoxidizer` executable which serves as a
CLI interface to performing common actions related to PyOxidizer.

We start by implementing an "analyze" command which can audit
executables for binary compatibility. It is currently limited to ELF
and is far from feature complete. But it is a start.