pyoxidizer: move pyrepackage crate into pyoxidizer crate
We need to expose pyoxidizer as a library as well as a binary so
that pyembed can continue to call into it. The goal is for the
`pyoxidizer` binary to take care of invoking the build-time logic
so downstream Rust crates don't need to pull in pyoxidizer directly
and only interact with its output.