bytecode: invoke Python process to compile bytecode
Compiling bytecode in process is the most efficient way to
compile bytecode. However, this required the binaries to link with
libpython. This required the host to have a linkable libpython
available. And, because Cargo doesn't support multiple crate
configurations from multiple crates, we had to create a copy of
the rust-cpython crate to facilitate compiling it with different
features.

This was all a giant hassle.

This commit removes the in-process code for compiling bytecode and
replaces it with a call to `python bytecodecompiler.py`. The
Python interpreter comes from the Python distribution we're
embedding. But it could theoretically come from anywhere. We
don't remotely support cross compiling yet, so we can cross this
bridge later.

After this commit, the pyoxidizer crate no longer needs the cpython
or python3-sys crates, so we remove their dependencies. And this
crate was the only referencing the copy of the rust-cpython
crate, so we remove those files from the repo. Good riddance.