projectmgmt: support for configuring a new PyOxidizer project
We want to make it easy to build PyOxidizer support into
binaries. This commit introduces high-level APIs and commands
for creating new projects that use PyOxidizer and for adding
PyOxidizer to existing projects. There is still a lot of work
to do and things still need to be documented. But this is a
start.

As part of this, the pyembed crate's files have been moved into
pyoxidizer. It looks as if it is a submodule now, but it isn't
actually used by the pyoxidizer crate as Rust source code. Instead,
the raw files are embedded in the binary so they can be installed
in target projects. Ideally it would exist as its own crate. But
getting the build integration to work is challenging. I think
this is the easiest approach for now.