pyrepackager: consolidate extension module rules into [[python_packages]]
Having unified processing for packaging rules seems better than having
extension modules live as a one-off. One of the benefits of including
extension modules with everything else is that we can now filter
extension modules via the "filter-file-include" rule, which will
facilitate pruning unused extension modules and further minimizing
binary size.

There's definitely room to improve how the rules are defined. For
now the logic is mostly a port of the old policy-based rules. Look
for tweaks in future commits...