pyinterp: support setting sys.path
Before, code for doing this was commented out. And we didn't have the
config file hooked up to the run-time config struct.

This commit implements everything.

I had trouble getting Py_SetPath() to work. So I just gave up and wrote
some code to mutate sys.path after initialization. The side-effects
should be mostly the same. There is probably some work to shore up the
various global variables influenced by a call to Py_SetPath() however.
We can worry about that another day.

I also changed behavior to use $ORIGIN as the token that expands to
the directory of the binary.