pyinterp: support handling SystemExit more gracefully
PyErr_PrintEx() may call exit() in the case of SystemExit. We don't
want this to occur because it mucks with Rust's lifetimes and prevents
our Drop implementation from being called.

This commit reimplements some low-level error handling in CPython to
support intercepting SystemExit and converting it to a Rust type so
that the Rust process is in control of calling exit().

As a result of this change, our Drop implementation should be called
and things are cleaned up as desired.

There's definitely room to improve the interpreter API to support
better control over error handling. But this is strictly better than
before.