pyembed: store references to source and bytecode data in module object
Before, we relied on global variables pointing to the module
source and bytecode data. After this commit, each Python module's
state contains references to these data structures. We still have
a global variable, but it only needs to be set for the narrow
window when initializing interpreters.

Theoretically this allows different interpreters to have different
module sets. Although I'm skeptical anybody will use this feature.
Flexibility is a sign of good design and implementation though.