pyembed: implement jemalloc raw allocator
The current Rust allocator has some overhead associated with it because we need
to parse Layout details and track their state. To be honest I'm not sure
if we need to track Layout state or not. Either way, there's no reason we
have to go through Rust's allocation layer: we can just call into jemalloc
directly.

This commit implements a Python raw allocator that speaks directly to jemalloc.
We retool the config mechanism so the allocator is now an enum with possible
values of jemalloc, rust, or system.

Jemalloc is the default allocator.