config: add "target" to Python distribution and allow defining multiple entries
I want to consolidate the TOML configs so multiple targets can share a
single config.

This commit converts [python_distribution] into [[python_distribution]]
and adds a "target" field to that section that denotes which Rust target
triple the distribution is for. We parse out the first section defining
a distribution matching the current current or error if we couldn't find
a suitable distribution.