importer: store references to importers in module state
We want to implement an uber importer so we only have 1 registered
meta path importer. In order to do this we need to store a reference to
the existing meta path importers. This commit implements that.

This required implementing a clear function for our module type
so we don't leak PyObject references.