importer: define a new unified importer class
The end goal is to have an instance of this class hold all meaningful
state and to use it as a single meta path importer. This commit
implements the basic scaffolding for a meta path importer and registers
our new type on sys.meta_path, replacing the builtin and frozen default
importers. Things appears to "just work" despite half the interface not
being implemented. We still need to integrate support for in-memory
importing. Future commits will do that. This commit is large enough as
it is.

Since state is now tracked on instances of the new type, we no longer
have to track various state in the module state, so that complexity goes
away.