main: add "build" command
This is a glorified wrapper around `cargo build`. Its utility is a
bit dubious at the moment. I want to improve things in the future by
adding a special build script mode that doesn't depend on the
pyoxidizer crate. Essentially, I want the "external" pyoxidizer to do
the heavy lifting and for the build script to be "dumb" so it
doesn't have to compile 100+ dependencies. We'll get there. You
have to start somewhere.