projectmgmt: call into `pyoxidizer` from build script
Previously, the build script for generated projects imported the
pyoxidizer crate to call a function within. This brought in 100+ crate
dependencies and made building very slow. This wasn't a good
experience and gave the illusion that generated binaries were
bloated.

This commit changes things around so that build script now invokes
the `pyoxidizer` executable and calls a command that runs the build
script code. Since build scripts communicate via special stdout
output, this should "just work."

The big risk to this change is that the invoking `pyoxidizer` may
not be the same version and therefore be incompatible with the
files in the project. This is an existing problem. We'll need to
add some version checking somewhere to shore this up. That can
be done later.

A regression from this change is the packaging output is no longer
written to stdout when running `pyoxidizer build`. We'll need to
come up with a solution to this. I imagine we'll pass a socket path
via an environment variable or something. There are options.

It turns out the 2 stage build introduced a few commits ago was
short lived. Fine: I never really liked that approach anyway.