pyrepackager: store resolved resources indexed by package
The importlib ResourceReader loads resources on a per-package basis. So,
resources need to be indexed by their package name.

Now that we've taught resource finding to annotate each resource with a
package name, it is possible to group resource files by their package.

This commit does that.

We still don't do anything with collected resource files, however.