repackage: support consuming pre-built artifacts
With this commit, pyembed's build.rs now looks for the
PYOXIDIZER_REUSE_ARTIFACTS and PYOXIDIZER_ARTIFACT_DIR
environment variables to engage a build mode where existing
artifacts on the filesystem are reused instead of calling into
PyOxidizer to perform packaging.

This will enable artifacts to be built before calling `cargo build`,
such as from `pyoxidizer build`.

One of the changes made by this commit is to introduce a compilation
environment variable holding the location of the generated data.rs
file. The file no longer must reside in OUT_DIR. This removes an
ugly hack to always copy the file there.