config: start to define install locations for packaging rules
The semantics are very tighly defined at the moment and we only
recognize the key on 1 rule type. But it's a start.

I fought Rust's borrow checker with iterators and lost. So I had to
stick a clone() in to work around it. It isn't performance sensitive, so
I'm not too worried.