repackage: define BuildContext to hold state for a build
We were carrying around a lot of variables related to the build
environment for an application and the logic was scattered. By
establishing a type to hold all this context, we can make things easier
to manage and grok.

There's still room to improve this type. For example, it doesn't hold
a reference to the config file or its parsed result. We'll get there.