repackage: initial support for app-relative install location
Sometimes a Python package doesn't work when embedded in a binary.
While this ideally wouldn't be the case, we can't live in an ivory
tower.

This commit implements initial support for installing resources into
paths relative to the produced application binary. There's still a
very long way to go and the feature is somewhat buggy. But this commit
represents a good stopping point for the implementation.