repackage: define build path in config file
As we start dabbling into installing files relative to the binary,
we're going to need to do more complicated things, such as copying
the produced binary to a staging area to live next to other installed
resources.

In preparation for this, we introduce a unified output directory
and a configuration setting to control it. This replaces our config
setting for setting the artifacts path.