repackage: define application name in build configuration
Before, we didn't know the name of the Rust application: we just
built everything and hoped the binary name aligned with the
name of the directory.

This commit establishes a required [[build]] application_name
config value that defines the name of the Rust binary corresponding
to our application. And since we now invoke `cargo build --bin`
with this value, we should now know the exact executable that is
being produced.