repackage: copy built binary into per-app directory
Previously, we built a binary in the Rust target directory and did
everything from there. This was somewhat constraining because there are
a ton of random files next to the binary.

This commit introduces the concept of a separate directory for staging
"packaged" application. Currently, the "packaging" phase simply copies
the produced executable to build/apps/<app_name>/<app_name>.