pyembed: configure PathFinder and sys.path during importer init
Previously, we registered PathFinder and updated sys.path after
Py_Initialize() returned. This caused problems when the standard
library wasn't embedded in the binary because Py_Initialize() needs
to load Python modules from the filesystem, which requires PathFinder
and sys.path to be configured.

Our solution to this problem is to configure these items as part of
initializing our custom module importer, therefore allowing them
to be set before Py_Initialize() finishes.

With this commit, we can now import the Python standard library from
the filesystem, just like normal Python.