repackage: serialize packaging state to disk
Previously, we processed our app-relative "packaging" rules at build time.
We want to make a clearer distinction between "build" and "packaging."

But, because they could be performed by separate processes, we need the
ability to exchange state between them.

This commit introduces the concept of "packaging state," which is a
data structure that can be serialized. We serialize the "packaging
state" to a CBOR file in the PyOxidizer artifacts directory. That
file/state is read at packaging time, if needed.

To close things out, we move the processing of app-relative resources to
the packaging stage.

There is still room to clean this up. For example, we may want a
function to resolve the configuration into a serializable data structure
without doing any I/O. Things can be cleaned up later.