projectmgmt: make jemalloc a configurable feature
Last night I disabled jemalloc by default because it didn't work
on Windows.

I didn't like the approach.

This commit refactors the jemalloc code yet again to be more flexible.

jemalloc is now configurable as a Rust crate feature. The feature
is disabled by default.

The [[embedded_python_config]] raw_allocator setting now influences
whether `--features jemalloc` is passed to `cargo build` when performing
a `pyoxidizer build`.

We change the default template for the written Cargo files and the
pyoxidizer.toml so the defaults "just work" on Windows and non-Windows.

Things still aren't perfect. We probably need an override to force
disable jemalloc or control the Cargo features to use. But I think
this is strictly better than what I did last night.