dist: pass explicit path to extract Python to
Before, we created a temporary directory in the function that extracted
a Python distribution archive.

After this commit, we create a temporary directory at a higher level and
pass the extraction path into the function. We also store the temporary
directory reference alongside the distribution metadata - not as part of
it - to prevent it from being deleted.

This will facilitate extracting Python distributions to persistent
paths.