distribution: initial WiX installer support
This commit teaches the distribution mechanism to produce
self-contained Windows installers using the WiX Toolkit.

The functionality is still somewhat limited. For example, we don't yet
support rendering the license text properly. We also don't support a
custom logo.

But the basic installer functionality is there. The generated .exe even
installs the Visual C++ Redistributable automatically, if needed.
Although it installs the full bundle redistributable instead of just
VC++ 2017. We could optimize this later, of course.