projectmgmt: ensure minimum Rust version requirements when building
Without this, the build could fail for reasons that may be confusing to
people who aren't Rust developers.

Let's add the minimum Rust version into the binary and verify it as part
of building applications.

There may be some special cases where we want to ignore this check.
Let's wait and see if anyone complains before adding complexity to skip
the check.

This should be the final component of fixing #18.