packaging_rule: use copy of distutils when invoking Python
In order to support building C extensions that are linked into a
single binary, we'll need to change how they are built. My plan for
doing this is to hack up distutils at build time so the C extension
building behavior is modified appropriately.

The first step of this is to get invocations of Python packaging tools
to use a version of distutils we control. This commit adds logic
to create a copy of distutils in a temporary directory and sets
PYTHONPATH so invoked Python processes pick it up.