distutils: vendor and use copies of distutils modules
Our strategy for hacking distutils will be to replace specific files
with modified versions wholesale. There are various trade-offs to other
approaches, such as maintaining patch files. But I'm a big fan of
determinism. This approach does tie PyOxidizer to a specific version
of Python, which is not great. But this is the easiest to implement. And
it seems appropriate for the current maturity level of PyOxidizer.

The distutils files were copied from CPython 3.7.3.