pyrepackager: support embedded C extension on UNIX platforms
The implementation is documented in the docs. This is still
somewhat half-baked. But infinitely better than what we had
before, which was nothing.

I haven't tested on macOS. It might "just work." Windows will
need to be implemented in a later commit, as it uses a different
distutils compiler backend.

I have successfully built and embedded extension modules for
zope.interface, zstandard, and Mercurial. So I'm pretty confident
that this works.