repackage: add libraries required by C extensions as link requirements
Basically we proxy the library requirements down to the linker for
the Rust binary. This means things should "just work." Although
there is the potential for the built binary to have additional
library dependencies, which may not be desired. We can deal with this
later (such as adding packaging rule functionality to deny shared
library linking).

I was able to build an executable containing the `pyyaml` package
and `import _pyyaml` with this change. So closing #66.