pyembed: always use exe path for Py_SetProgramName()
There were a few problems with the previous use of Py_SetProgramName().
First, the path should be absolute, not relative. Before we were
passing a relative path. Second, on Windows the path did not include
.exe.

Since Py_SetProgramName() is supposed to be the value of argv[0], I
don't think it makes sense to pass in any value that isn't argv[0].
So, this commit removes the option for specifying an explicit value
and always passes in the current executable's path.

A side-effect of this change is that `import ctypes` now works on
Windows! This is because our custom builds of Python tell ctypes
to look for Python symbols in `sys.executable`. Previously,
`sys.executable` referred to an invalid filesystem path because
there was no `.exe` extension. With `sys.executable` being properly
populated after this commit, `import ctypes` works! This thus
closes #61.