wix: build multiple architecture installer
Now that we support building on 32-bit Windows, we can teach WiX to
generate 32-bit MSI installers.

As part of this, we teach WiX to build a multiple architecture .exe
installer. The .exe will run on 32-bit and 64-bit Windows. At run-time
it will detect whether we're on 64-bit Windows and will install the
appropriate application version. This means application maintainers
only have to provide a single installer to end users.