packaging_rule: pass possible install paths to PYTHONPATH
setuptools insists that the path we're installing to appears on
sys.path. In order to appease setuptools, this commit sets
PYTHONPATH to contain the paths that we could install packages to.

This is a partial fix for #70 and eliminates the error reported
in that issue. But another setuptools issue remains...