docs: Update sample `pyoxidizer run` output in first example.
This is what it looks like for me with the v0.2.0 release (modulo
adjusting the project's root directory to follow the example.)

Various other bits of output in examples have also changed.  But
it'd be a lot of work to update them all comprehensively, and I
think better to update piecemeal than not at all.