docs: Be explicit up front about getting the actual executable.
Looking at the getting-started page of the docs, because everything
is in terms of `pyoxidizer run` it's not obvious how to actually
lay hands on the single executable file that PyOxidizer produces.
Here's one example of a puzzled would-be user:
  https://groups.google.com/forum/#!topic/pyoxidizer-users/rUyM3iTpZq0

Fix that by breaking up the first example into `pyoxidizer build`
and then an explicit invocation of the built executable.