pyembed: fix InspectLoader.get_source() incompatibility with memoryview
The function we were calling does not properly handle memoryview
instances. So we need to cast to bytes first in order for things
to work.

With this change, `loader.get_source(modname)` works with the
in-memory importer!

With this change, I am able to `import pkg_resources` from memory.
So this fixes #134.