pyrepackager: write .pyc files for app-relative modules
This fulfills some TODOs in the code.

The behavior is slightly different from Python in that we do not
check source hashes when loading bytecode. This is because app-relative
files should be immutable.

This assumption may be invalid in some corner cases for some
applications. But it seems like a reasonable default for now.