pyrepackager: correctly create app-relative packages
We track whether a module is a package for a reason. Use it
to determine whether to create a __init__ file instead of sniffing
the known set of packages.

This should enable single file packages to work correctly.

I discovered this when packaging Mercurial. Its mercurial.cext
package only contains an __init__.py when the C extensions are
embedded in the binary. mercurial.cext wasn't being detected
as a package using name sniffing. But the filesystem scanning
had properly flagged it as a package.