starlark: define Config to represent a full configuration
We also had to add the config path to the environment so all available
fields are present.

With this commit, we can now represent a full "parsed" configuration
completely via Starlark!

The type conversion for packaging rules in this patch is objectively
horrible. I plan to refactor all this stuff soon. But first we
need to get Starlark working.