global: require Rust 1.36+
We need this in order to support the starlark crate, which we will
soon introduce.

There is a dependency crate that fails to compile on 1.31. But
starlark itself is the only blocker on 1.32-1.35. And the starlark
failure looks pretty trivial (just a minor lifetime issue). So
we could potentially restore support for 1.32+ relatively easily
by making some minor changes upstream.