rust-cpython: synchronize crate with submitted changes
I've submitted pull requests to rust-cpython to incorporate all
the changes we needed to make to support PyOxidizer. This commit
synchronizes our vendored crate with the final submission state
of all those changes so that we can more accurately test the
proposed upstream changes.

Hopefully this is the last commit to the vendored rust-cpython
crate before we delete it and use a vanilla version published
by upstream...