starlark: manage build path as part of EnvironmentContext
In preparation for allowing the context to track more state,
such as the directory for Python distribution archives.

As part of this, we change the Starlark API so we use explicit
functions to change the build path instead of a variable. While
I did this so we could more easily intercept functionality, this
feels like a better API design.