starlark: initial implementation of FileManifest
We start to implement support for FileManifest in Starlark to
represent a virtual manifest of files. We start by teaching the
type how to add some Python resource types to instances.

This is far from feature complete. But you need to start somewhere.
Perhaps the biggest limitation here is the lack of support for
bytecode and extension module files. Those are just stubs for
now. To implement these will require plumbing in a reference to
the Python distribution somewhere so we can compile bytecode
and link extension modules.