starlark: implement PythonDistribution.pip_install()
This commit demonstrates more of the future of the packaging rules.
We can now run pip from a PythonDistribution instance and directly
collect the resources it installs. From there, we can inspect,
filter, and add those resources however we see fit using the
power of Starlark. This is vastly more flexible than the previous
approach of adding arguments to control every nuanced behavior.

We still need to implement a ton of APIs in Starlark to make the
new packaging module usable. But this is a start...