starlark: remove PipInstallSimple and PipRequirementsFile rules
These are now implemented as methods on the Python distribution.

It is possible a bit too premature to delete these, as we're still
missing functionality in Starlark. But I don't plan on having a
release until Starlark is feature complete and it doesn't make sense
to have multiple versions of this logic in the repository. Let's
nuke the code and force all development to occur within Starlark.