py_packaging: add API for building executable from Rust project
We're slowly reimplementing logic from repackage.rs and projectmgmt.rs
in a more flexible manner. This commit introduces a new API to build
a PreBuiltPythonExecutable using a Rust project.

As part of implementing this, we discovered a few bugs in the new code.
The library path wasn't being emitted to Cargo and we weren't
capturing the libpyembeddedconfig library data, leading to a linker
error. We've taught the linking and building mechanism about the missing
pieces.

Even with this in place, linking still fails due to missing symbols.
I think the probably is certain object files aren't being included. This
could be the result of using an empty embeddable resources object. This
will need closer examination. For now, the test is coded to assume cargo
failure.