starlark: build executable when adding PythonExecutable to a FileManifest
Now that we have more usable APIs for building binaries/projects, it is
finally time to hook it up to Starlark.

This commit teaches FileManifest.add_python_resource() to build an
executable from PythonExecutable instances. Under the hood, a temporary
Rust project is created and built and the resulting executable is added
to the FileManifest.

In order to preserve the add_python_resource() API, we added a "name"
argument to PythonExecutable(). There were also a handful of minor
changes made to make this all work.