app_packaging: use system allocator by default in tests
This commit is a bit sketchy. But as we add tests that build Rust
projects, build times become a concern and jemalloc slows things
down considerably. Disabling jemalloc is a quick way to speed up
the builds. Not using jemalloc as the allocator when testing feels
like a quick and dirty win.