starlark: remove clone from ParsedPythonDistribution
This type is large and we shouldn't be cloning it. In order to remove
clone(), we had to convert references to this type to Arc<T>. Strictly
speaking, we probably could have used Rc<T>. But we needed an Arc<T> to
compile in test mode and I didn't want to go down the rabbit hole.
Starlark types use Arc<T> anyway, so there's a good chance we may need
this someday anyway.