app_packaging: use new APIs for producing files
Most of the code in process_config() has been reimplemented elsewhere.
This commit ports that function to use the newer code.

As part of this, I also nuked PackagingState. I intended to do this as
part of a separate commit but forgot to actually commit. Oops. This
was only used to write license files. While license files aren't yet
ported to the new API, I don't want to keep this code lingering any
longer, as it inhibits further refactoring.

There is still a lot of mostly dead code to be nuked. But I want to do
this incrementally so I'm more confident that all functionality
exists in the new code paths...

It's worth noting that the executables built with the legacy code paths
have been busted ever since we ripped out the legacy packaging rules.
That's because the set of embedded Python modules is effectively
empty since there are no packaging rules to populate them. The only
functionality that still works is `cargo build` doesn't error.