starlark: remove Config
This functionality hasn't been used for a while and was the last
holdover of the legacy config mechanism. Let's nuke it.

In order to get this to compile, we had to hack up the Config struct
to remove an annoying-to-populate field. The affected code paths
shouldn't be used and will be deleted soon enough anyway, hopefully.