starlark: introduce concept of registered build targets
We want a single configuration file to be able to define multiple
actions. And we want it to be possible to run one action in isolation
from others.

In order to facilitate this, this commit introduces the concept of
"targets" to configuration files.

Configuration files now register functions which define build targets
that can be resolved. While not yet implemented, at evaluation time,
the configuration file will be given a list of targets to evaluate
and each target's function will be called accordingly.

We implement a new list-targets command to query the available build
targets.