starlark: add resolve_targets()
This commit adds a function to resolve targets via their registered
build functions. Arguably this shouldn't be necessary and we could
evaluate some Starlark statements outside of Starlark. But I couldn't
figure out how to author Rust code using the available APIs to do this.
This commit is me giving up on the idea for the time being so we can
make forward progress.

We still need to pass in the requested targets properly. That will be
done in a future commit.