starlark: start to implement generic running functionality
We want targets in Starlark to be "runnable." That way we can
provide a `pyoxidizer run` command that will run whatever we built.

This commit starts that process by introducing a trait that
indicates a starlark type represents a "build target" and can
be runnable. We only implement this for FileManifest so far.
And the implementation is not yet implemented. But the dynamic
dispatch does appear to work and the `pyoxidizer run` command
does invoke the new functionality. So it is a start. Future
commits will actually implement useful functionality.