starlark: implement BuildTarget for PreBuildPythonExecutable
This will allow us to transform this type into a build executable, which
can be run.

Again, we haven't implemented it in a meaningful way yet. In this case,
we need to pass more context into the build() function. So we'll have to
invent a type for that and modify the trait to accept it as an argument.