starlark: pass context struct to build()
build() implementations may need to know things like where to write
files. This commit introduces a struct for passing along this
data.

The first piece of data we pass is the path where we should write
files.

We still do not have a meaningful implementation of any build().
That will come soon.