starlark: implement build() for PreBuiltPythonExecutable
It now creates a temporary Rust project for the executable and builds
it, writing the executable to the output directory.

For some reason, the executable isn't actually executable. So we
get a permissions error when attempting to execute it via
`pyoxidizer run`. This will be fixed in a subsequent commit.