projectmgmt: allow specifying target to run
This required a new argument to the command to specify a non-default
target.

After this commit, we clearly see `pyoxidizer run --target X` resolving
the specified target and `pyoxidizer run --target install` works with
the default configuration file!

At this point, our Starlark functionality is now more or less on
par with what we had before in terms of project building and
running functionality. There are still some features to reimplement
such as caching of build artifacts. But the high-level commands
are working and everything is driven by Starlark.