starlark: declare dependencies for targets
Starlark doesn't allow recursive function calls. So, we can't call
`resolve_target()` from a function that is itself a target function
being called by `resolve_target()`.

Our solution for this is to declare target dependencies as part of
`register_target()`. The registered targets now record a list of
targets they depend on and the resolved value for each target is passed
as an argument to the target function.

In order to placate Rust's insistence that there be a single mutable
borrow at a time, we needed to refactor the logic for resolving
targets to carefully avoid overlapping borrow scopes.

Because it is weird to reference a target in the depends list before
it is itself registered as a target, we added a `default` argument to
`register_target()` to explicitly mark a target as the default target.