starlark: move embedded resources type into PythonExecutable
While the struct exists under the hood, I think it makes sense to not
expose it to Starlark, as it is an extra moving part and doesn't add
any value being at present being a standalone entity. We may want to
reconsider this once we have multiple entities that we'll want to
add resources to, as its existence could cut down on some boilerplate.