starlark: move construct of PythonExecutable to a PythonDistribution method
When types are used to bootstrap other types, I think it makes sense
for that transformation to live as a method of the source type.

With this change, we also introduce a new approach to Starlark where
we implement the bulk of the logic outside a macro. I think this pattern
will improve maintainability, as the code looks like normal Rust and is
easier to read.

The code would look even nicer if we didn't have a hacky code path to
write build artifacts here. I plan to clean this up, as I'm pretty sure
the functionality it supports is busted.