starlark: move PythonDistribution() into method
And with this all of the methods for the Starlark type are implemented
outside of macros and the in-macro code simply calls out to a method.

The new code is vastly easier to read because it is formatted properly
and IDE type inference works correctly. The only loss of readability
is with default argument values, which are defined in the starlark
macro, away from the methods which accept &Value. That's an acceptable
price to pay.